LayoutTests/imported/w3c/web-platform-tests/fetch/api/policies/referrer-unsafe-url-expected.txt

CONSOLE MESSAGE: Unrecognized Content-Security-Policy directive 'referrer'.


PASS Request's referrer is the full url of current document/worker 
